+++ tlm_cmd/telemetry_generator.c
@@ -11,7 +11,7 @@
 #include <src_user/tlm_cmd/telemetry_definitions.h>
 #include <src_user/tlm_cmd/user_packet_handler.h>
 #include <src_user/settings/tlm_cmd/common_tlm_packet_define.h>
-#include "./ccsds/tlm_space_packet.h"   // FIXME: TSP 依存性はNGなので， TCP → SP 大工事終わったら直す
+#include "./ccsds/space_packet_protocol/tlm_space_packet.h"   // FIXME: TSP 依存性はNGなので， TCP → SP 大工事終わったら直す
 
 
 /**